+++ package.json
@@ -4,7 +4,7 @@
   "description": "A web app that uses Cheerio and MongoDB to scrape news stories and allows users to comment on them.",
   "main": "server.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
   },
   "repository": {
     "type": "git",
